+++ src/main/java/com/eliteessentials/integration/EliteEssentialsEconomy.java
@@ -241,7 +241,7 @@ public class EliteEssentialsEconomy implements Economy {
             );
         }
         
-        boolean success = playerService.removeMoney(accountID, amount.doubleValue(), null);
+        boolean success = playerService.removeMoney(accountID, amount.doubleValue(), null, null);
         EconomyResponse.ResponseType status = success ? 
             EconomyResponse.ResponseType.SUCCESS : 
             EconomyResponse.ResponseType.FAILURE;
@@ -278,7 +278,7 @@ public class EliteEssentialsEconomy implements Economy {
             );
         }
         
-        boolean success = playerService.addMoney(accountID, amount.doubleValue(), null);
+        boolean success = playerService.addMoney(accountID, amount.doubleValue(), null, null);
         EconomyResponse.ResponseType status = success ? 
             EconomyResponse.ResponseType.SUCCESS : 
             EconomyResponse.ResponseType.FAILURE;